+++ src/pke/unittest/UnitTestSerializePKE.cpp
@@ -225,7 +225,7 @@ TEST_F(UTPKESer, Keys_and_ciphertext) {
 		EXPECT_EQ( *kp.secretKey, *kpnew.secretKey ) << "Secret key mismatch after ser/deser";
 	}
 	DEBUG("step 3");
-	vector<uint32_t> vals = { 1,3,5,7,9,2,4,6,8,11 };
+	vector<uint64_t> vals = { 1,3,5,7,9,2,4,6,8,11 };
 	Plaintext plaintextShort = cc->MakeCoefPackedPlaintext( vals );
 	Ciphertext<Poly> ciphertext = cc->Encrypt(kp.publicKey, plaintextShort);
